fix(tests): extract auth-poll and credential-store-impl to isolate mock.module leakage
Bun's mock.module() is process-global and retroactive — it replaces live
ESM bindings even for static imports. login-flow.test.ts mocks both
"auth" and "credential-store", causing auth-poll.test.ts and
credential-store.test.ts to receive mocked implementations instead of
real ones when running in the same Bun process.

Fix: extract pollForAccessToken into src/helpers/auth-poll.ts and the
credential-store implementation into src/helpers/credential-store-impl.ts.
Tests import from these new paths which are not targeted by mock.module().
The original modules become thin re-export wrappers for backward compat.

Also ensure all mock.module() calls provide complete export surfaces so
any test file that resolves a mocked module gets callable functions
instead of undefined.
