+++ src/helpers/auth.ts
@@ -11,13 +11,18 @@
 import { logDebug } from "./log";
 import { SignupRequiredError, isSignupRequiredError } from "./signup";
 
+// Re-export pollForAccessToken from its own module so that existing consumers
+// (login-flow.ts, etc.) keep working with `import { pollForAccessToken } from "./auth"`.
+// The separate file exists to isolate the real implementation from mock.module()
+// calls in login-flow.test.ts — see auth-poll.ts header for details.
+export { pollForAccessToken } from "./auth-poll";
+
 // ---------------------------------------------------------------------------
 // Constants
 // ---------------------------------------------------------------------------
 
 const PLUGINS_API = "https://plugins.archgate.dev";
 const GITHUB_DEVICE_CODE_URL = "https://github.com/login/device/code";
-const GITHUB_DEVICE_TOKEN_URL = "https://github.com/login/oauth/access_token";
 
 /**
  * GitHub OAuth App client ID for the archgate CLI (public client — no secret).
@@ -36,27 +41,6 @@ interface DeviceCodeResponse {
   expires_in: number;
   interval: number;
 }
-
-interface DeviceTokenSuccessResponse {
-  access_token: string;
-  token_type: string;
-  scope: string;
-}
-
-interface DeviceTokenPendingResponse {
-  error: "authorization_pending" | "slow_down";
-  error_description?: string;
-}
-
-interface DeviceTokenErrorResponse {
-  error: "expired_token" | "access_denied" | string;
-  error_description?: string;
-}
-
-type DeviceTokenResponse =
-  | DeviceTokenSuccessResponse
-  | DeviceTokenPendingResponse
-  | DeviceTokenErrorResponse;
 
 // ---------------------------------------------------------------------------
 // GitHub Device Flow
@@ -85,72 +69,6 @@ export async function requestDeviceCode(): Promise<DeviceCodeResponse> {
   const data = (await response.json()) as DeviceCodeResponse;
   logDebug("Device code received, expires in:", data.expires_in, "seconds");
   return data;
-}
-
-/**
- * Step 2: Poll GitHub until the user authorizes (or the code expires).
- * Returns the GitHub access token on success.
- */
-export async function pollForAccessToken(
-  deviceCode: string,
-  interval: number,
-  expiresIn: number
-): Promise<string> {
-  const deadline = Date.now() + expiresIn * 1000;
-  let pollInterval = interval;
-  logDebug(
-    "Starting device code polling, deadline:",
-    new Date(deadline).toISOString()
-  );
-
-  /* oxlint-disable no-await-in-loop -- sequential polling is required by RFC 8628 */
-  while (Date.now() < deadline) {
-    await Bun.sleep(pollInterval * 1000);
-
-    const response = await fetch(GITHUB_DEVICE_TOKEN_URL, {
-      method: "POST",
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        client_id: GITHUB_CLIENT_ID,
-        device_code: deviceCode,
-        grant_type: "urn:ietf:params:oauth:grant-type:device_code",
-      }),
-      signal: AbortSignal.timeout(15_000),
-    });
-
-    if (!response.ok) {
-      throw new Error(`GitHub token poll failed (HTTP ${response.status})`);
-    }
-
-    const data = (await response.json()) as DeviceTokenResponse;
-
-    if ("access_token" in data) {
-      logDebug("Access token received");
-      return data.access_token;
-    }
-
-    if ("error" in data) {
-      if (data.error === "authorization_pending") {
-        logDebug("Authorization pending, retrying in", pollInterval, "seconds");
-        continue;
-      }
-      if (data.error === "slow_down") {
-        pollInterval += 5;
-        logDebug("Slow down requested, new interval:", pollInterval, "seconds");
-        continue;
-      }
-      // expired_token, access_denied, or other terminal error
-      throw new Error(
-        data.error_description ?? `GitHub authorization failed: ${data.error}`
-      );
-    }
-  }
-  /* oxlint-enable no-await-in-loop */
-
-  throw new Error("Device code expired. Please try again.");
 }
 
 interface GitHubUserInfo {
